Arduino: Add MKRWANFWUpdate_standalone sketch - ARD-078 1.2.3
Add the sketch from MKRWAN library examples for revision control. Compile and
install and run this sketch to upgrade the MKRWAN1310 LoRa Firmware to the
version:

  ARD-078 1.2.3

This sketch will run and reinstall the firmware whenever the MKRWAN1310 board is
powered on, so replace with another sketch once firmware upgrade has been done.